collect and pickle reddit_tifu data
The data is collected from tensorflow_datasets. The dataset is first stored into the folder /data/reddit_tifu

Afterwards the dataset is preprocessed by selecting the relevant columns and decoding them to UTF-8.

Lastly the resulting dataframe is pickled so the retrieval process doesn't have to be repeated on each run.